requirements.txt
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
flask
flask-cors
Pillow
rembg==2.0.55